LPS-115019 Move to its own module because SEarch Container is not a Clay TagLib
