+++ modules/apps/frontend-taglib/frontend-taglib-search-container-view-state/src/main/java/com/liferay/frontend/taglib/search/container/view/state/SearchContainerViewState.java
@@ -12,9 +12,7 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.clay.servlet.taglib.view.state;
-
-import javax.portlet.RenderRequest;
+package com.liferay.frontend.taglib.search.container.view.state;
 
 import org.osgi.annotation.versioning.ProviderType;
 
@@ -22,15 +20,24 @@ import org.osgi.annotation.versioning.ProviderType;
  * @author  Neil Griffin
  */
 @ProviderType
-public interface SearchContainerViewStateFactory {
+public interface SearchContainerViewState {
 
-	public SearchContainerViewState create(
-		int cur, int delta, String displayStyle, int end, String keywords,
-		String orderByCol, String orderByType, boolean resetCur, int start);
+	public int getCur();
 
-	public SearchContainerViewState create(
-		String defaultDisplayStyle, String defaultOrderByCol,
-		String defaultOrderByType, RenderRequest renderRequest,
-		String[] validOrderByCols);
+	public int getDelta();
+
+	public String getDisplayStyle();
+
+	public int getEnd();
+
+	public String getKeywords();
+
+	public String getOrderByCol();
+
+	public String getOrderByType();
+
+	public boolean getResetCur();
+
+	public int getStart();
 
 }
