+++ modules/apps/frontend-taglib/frontend-taglib-search-container-view-state/src/main/java/com/liferay/frontend/taglib/search/container/view/state/SearchContainerViewStateWrapper.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.clay.servlet.taglib.view.state;
+package com.liferay.frontend.taglib.search.container.view.state;
 
 import org.osgi.annotation.versioning.ConsumerType;
