+++ modules/apps/frontend-taglib/frontend-taglib-search-container-view-state/src/main/java/com/liferay/frontend/taglib/search/container/view/state/internal/SearchContainerViewStateFactoryImpl.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.clay.internal.servlet.taglib.view.state;
+package com.liferay.frontend.taglib.search.container.view.state.internal;
 
-import com.liferay.frontend.taglib.clay.servlet.taglib.view.state.SearchContainerViewState;
-import com.liferay.frontend.taglib.clay.servlet.taglib.view.state.SearchContainerViewStateFactory;
+import com.liferay.frontend.taglib.search.container.view.state.SearchContainerViewState;
+import com.liferay.frontend.taglib.search.container.view.state.SearchContainerViewStateFactory;
 import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.util.GetterUtil;
