+++ modules/apps/frontend-taglib/frontend-taglib-search-container-view-state/src/main/resources/com/liferay/frontend/taglib/search/container/view/state/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
